fixed moving components
