+++ ClientApp/src/components/MainPage.jsx
@@ -10,7 +10,7 @@ import { Messages } from '../pages/XboxAccountInfoPages/Messages'
 import { XboxOneGames } from '../pages/XboxAccountInfoPages/XboxOneGames'
 import { About } from '../pages/About'
 import { UserSettings } from '../pages/Settings/UserSettings'
-import { UpdateUser } from '../pages/UpdateUser'
+import { UpdateUser } from '../pages/Settings/UpdateUser'
 import { AddAccount } from '../pages/Settings/AddAccount'
 import { ViewAccounts } from '../pages/Settings/ViewAccounts'
 import { getUser, getFirstAccount } from '../auth'
